server/pyproject.toml
[project]
name = "todolist-server"
version = "0.1.0"
description = "TodoList API service"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "python-multipart>=0.0.9",
    "uvicorn[standard]==0.32.1",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "pydantic-settings==2.6.1",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",
    "requests>=2.32.0",
    "PyJWT>=2.9.0",
    "bcrypt>=4.2.0",
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.14.0",
    "deepagents>=0.1",
    "langchain-openai>=0.3",
    "tavily-python>=0.5",
    "langgraph-checkpoint-postgres>=2.0",
    "langgraph>=0.4",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
whisper = [
    "faster-whisper==1.1.1",
]
fun-asr = [
    "dashscope>=1.23.1",
]
dev = [
    "faster-whisper==1.1.1",
    "dashscope>=1.23.1",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
    "httpx==0.28.1",
    "ruff>=0.9",
    "pre-commit>=4.1",
]

[tool.uv]
# onnxruntime 1.24+ 不再发布 macOS x86_64 wheel；纳入解析环境后 uv 会选用仍带 Intel Mac 轮子的 1.23.x。
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "T20",  # flake8-print (catches stray print/pprint)
    "S",    # flake8-bandit (security)
    "C4",   # flake8-comprehensions
    "RET",  # flake8-return
    "ARG",  # flake8-unused-arguments
]
ignore = [
    "S101",  # assert is fine in tests
    "B008",  # Depends() in FastAPI default args is idiomatic
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "ARG", "S105", "S106"]
"alembic/**" = ["E501"]
"scripts/**" = ["T20"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
